reddit_analyzer/requirements.txt
praw
openai
firebase-admin
python-dotenv
requests
